test(ci): consolidate checkout fixture clocks
